(fix) recover jmh code
