+++ jraft-rheakv/rheakv-core/src/test/java/com/alipay/sofa/jraft/rhea/util/VarIntsBenchmark.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.alipay.sofa.jraft.benchmark.common;
+package com.alipay.sofa.jraft.rhea.util;
 
 import java.util.concurrent.TimeUnit;
 
@@ -29,8 +29,6 @@ import org.openjdk.jmh.runner.RunnerException;
 import org.openjdk.jmh.runner.options.Options;
 import org.openjdk.jmh.runner.options.OptionsBuilder;
 import org.openjdk.jmh.runner.options.TimeValue;
-
-import com.alipay.sofa.jraft.rhea.util.VarInts;
 
 /**
  * @author jiachun.fjc
